+++ docs/changelog.md
@@ -10,6 +10,10 @@ Each revision is versioned by the date of the revision.
 
 ## 2026-01-15
 
+- Added header rewrite support for gRPC backends.
+
+## 2026-01-15
+
 - Added missing settings from haproxy-route-tcp relation template.
 
 ## 2026-01-09